requirements.txt
kafka-python==2.0.2
minio==7.2.3
pandas==2.2.1
pyarrow==15.0.0
duckdb==0.10.2
